Using uint cause compilation error on Windows(WSL)

+++ src/validation.cpp
@@ -1116,7 +1116,7 @@ bool ReadBlockFromDisk(CBlock& block, const CDiskBlockPos& pos, int nHeight)
 bool ReadBlockFromDisk(CBlock& block, const CBlockIndex* pindex)
 {
     CDiskBlockPos blockPos;
-    uint height = 0;
+    uint64_t height = 0;
     {
         LOCK(cs_main);
         blockPos = pindex->GetBlockPos();
@@ -2964,7 +2964,7 @@ bool CheckBlock(const CBlock& block, CValidationState& state, bool fCheckPOW, bo
 
     // Check that the header is valid (particularly PoW).  This is mostly
     // redundant with the call in AcceptBlockHeader.
-    uint height = block.vtx[0]->vin[0].prevout.n;
+    uint64_t height = block.vtx[0]->vin[0].prevout.n;
 
     if (!CheckBlockHeader(block, state, height, fCheckPOW))
         return false;
